# src/api/v1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "this is a email send API with failover capabilities",
    "version": "1.0.0",
    "title": "service-failover-api",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "pavan8sachi@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api/v1",
  "tags": [
    
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/status": {
      "get": {
        "tags": [
 
        ],
        "summary": "check health of application",
        "operationId": "status",
        "produces": [
          "text/plain"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "text"
            }
          },
          "500": {
            "description": "unhealthy"
          }
        }
      }
    },
    "/messages": {
      "post": {
        "tags": [
         
        ],
        "summary": "check health of application",
        "operationId": "messages",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "email defintion to send",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MailRequest"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "successful operation",
            "schema": {
              "type": "object"
            },
            "examples": {
              "application/json": {
                "message": "mail has been successfully sent"
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "error operation",
            "examples": {
              "application/json": {
                "message": "sending mail has failed"
              }
            }
          }
        }
      }
    }
  },
  "securityDefinitions": {
   
  },
  "definitions": {
    "MailRequest": {
      "type": "object",
      "required": [
        "from",
        "to"
      ],
      "properties": {
        "mail_from": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "subject": {
          "type": "string"
        },
        "mail_to": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "mail_cc": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "mail_bcc": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}